CHROMIUM: iio: cros_ec: Enable activity events with sensor_ring enable
When sensor ring is enabled, all events go to cros_ec_sensor_ring.
However, for the activity sensor and powerd to work, activity events
should be send to cros_ec_activity driver as well.

This change is a stopgap solution and will be removed once
cros_ec_sensor_ring is deprecated when USE=iioservice is compile.

BUG=b:123434029
TEST=Compile with both USE=iioservice set or not.

Change-Id: I7b5e2806552646870da023490da36f0724d27474
Signed-off-by: Gwendal Grignou <gwendal@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2654329
Reviewed-by: Ching-Kang Yen <chingkang@chromium.org>
Tested-by: Ching-Kang Yen <chingkang@chromium.org>