CHROMIUM: drm/virtio: implement context init: track valid capabilities in a mask
Backport of commit dbf00e9759cd ("CHROMIUM: drm/virtio: implement
context init: track valid capabilities in a mask") from v5.10.

The valid capability IDs are between 1 to 31, and defined in the
virtio gpu spec.  This is used for error checking the subsequent
patches.  We're currently only using 2 capability IDs, so this
should be plenty for the immediate future.

BUG=b:176158390
BUG=b:178994481
TEST=glxgears

Signed-off-by: Ryan Neph <ryanneph@google.com>
Change-Id: I8d38a80423bcf910f0042cc7fe59621475d74409
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2669917
Reviewed-by: Gurchetan Singh <gurchetansingh@chromium.org>