CHROMIUM: power: supply: pchg: Handle PCHG_STATE_FULL received from CrOS EC
Currently, CrOS EC sends PCHG_STATE_DETECTED together with the
battery percentage == 100. This combination was intended to be
used as the indication of 'full' status.

However, peripheral device chargers usually stop charging when
the battery level is close to but not equal to 100%. Since the
thresholds differ chip to chip, it's better to leave the decision
in the EC.

This patch makes cros-pd-pchg set the PCHG status to 'Full' when
PCHG_STATE_FULL is received from CrOS EC.

Since the EC sets the battery percentage at intended timings, the
PCHG driver shouldn't manipulate it and add unnecessary complexity.

This patch makes the PCHG driver take the battery percentage
reported from CrOS EC as-is in all cases.

BUG=b:179390065,b:174310207
TEST=Verify /sys/class/power_supply/PCHG0/status is set to Full.
when stylus is fully charged.
TEST=Verify /sys/class/power_supply/PCHG0/capacity changes as
expected when stylus is charged, attached, and detached.

Signed-off-by: Daisuke Nojiri <dnojiri@chromium.org>
Change-Id: Ib0273deb02c008f4578b92e179f6c4507cb6afe5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2691446
Reviewed-by: Guenter Roeck <groeck@chromium.org>