CHROMIUM: thunderbolt: Detect all cases when no devices are attached
When a charger device is connected in reverse,
USB_PD_MUX_POLARITY_INVERTED bit is set in mux mode.When checking
for cases where no device attached, include this check as well.

BUG=b:175045931
TEST=Ensure enumeration of retimers with charger connected
normally and in reverse work fine.

TOREVERT=Once the upstream version of the patch lands in
maintainers tree, this patch should be reverted.
TOREVERT=b:180449562

Change-Id: Iceb6f555c958b84be282f77e2b3c4d7ba1bcbad7
Signed-off-by: Rajmohan Mani <rajmohan.mani@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2660600
Tested-by: Kranthi Kuntala <kranthi.kuntala@intel.corp-partner.google.com>
Reviewed-by: Prashant Malani <pmalani@chromium.org>