CHROMIUM: thunderbolt: Defer scan of retimers when no devices are connected
When no devices are connected, retimers can still be enumerated to
enable operations such as retimer firmware update. Defer the scan of
retimers to a later time, when no devices are connected. This gives
enough time for all the drivers to load, so retimer communication
can be successfully enabled.

BUG=b:175045931
TEST=Boot with no devices attached to Type-C port and update retimer
firmware on a Type-C port and ensure Type-C based devices work fine
on all Type-C ports.

TOREVERT=Once the upstream version of the patch lands in
maintainers tree, this patch should be reverted.
TOREVERT=b:180449562

Change-Id: I8c6a44568a5bd59814754a3f03f075cd90add0c1
Signed-off-by: Nikunj A. Dadhania <nikunj.dadhania@intel.com>
Signed-off-by: Rajmohan Mani <rajmohan.mani@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2670719
Reviewed-by: Prashant Malani <pmalani@chromium.org>
Reviewed-by: Alex Levin <levinale@google.com>