CHROMIUM: arm64: dts: qcom: sc7180-trogdor: Reorganize fingerprint for coachz-rev2
Applied the FPMCU_SEL -> EN_FP_RAILS change, since this will be
required on rev2. On rev1 this repurposing is a noop as FPMCU_SEL goes nowhere,
rev0 will break, but nobody expects that to work.

Removed the pinctrl and pin{mux,conf} for the control pins because:
1. The only need for them is for userspace control via flash_fp_mcu
2. cros-ec doesn't know what to do with them, and even if it did,
   it would interfere with flash_fp_mcu at the most inopportune times

Since we're not using hogs, we rely on AP firmware to set all the
control pins correctly, see chromium:2698272.

TEST=Boot on coachz-rev1, fingerprint should probe from boot
TEST=flash_fp_mcu_coachz should work
BUG=b:170284663, b:174803639, b:179530529

Cq-Depend: chromium:2698272
Signed-off-by: Alexandru M Stan <amstan@chromium.org>
Change-Id: Ifd7b86f826b18410eada75758a7bca1eebfa336d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2682978
Reviewed-by: Stephen Boyd <swboyd@chromium.org>
Reviewed-by: Craig Hesling <hesling@chromium.org>