CHROMIUM: iwl7000: support simulated PCI devices
The extra thing we need them to do (in the case of USIM, NOT ESL)
is to support the request_firmware callback so that we don't have
to have the file on disk matching the TLVs USIM was built with.

To do that, the simulated device exposes a second BAR (BAR 2, not
1, since we have 64-bit BARs) that can be accessed in magic ways
by doing readl() for the size and memcpy_fromio() for the data at
a given offset (now only 0 for the TLV image is supported, but we
may want to add support for other things.)

If this second BAR isn't supported we fall back to normal image
load from disk, which is the case for real hardware and thus also
ESL - that way, we can support both at the same time.

Change-Id: I44170643a91957d1d74d88c0ff49f7b9402c4955
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
iwl7000-tree: 5735cdbb09ae3945bbb166ed4dd174d5a8f655de