FROMGIT: xhci: fix potential array out of bounds with several interrupters
The Max Interrupters supported by the controller is given in a 10bit
wide bitfield, but the driver uses a fixed 128 size array to index these
interrupters.

Klockwork reports a possible array out of bounds case which in theory
is possible. In practice this hasn't been hit as a common number of Max
Interrupters for new controllers is 8, not even close to 128.

This needs to be fixed anyway

Signed-off-by: Mathias Nyman <mathias.nyman@linux.intel.com>
(cherry picked from commit baa4e8184909b4cdb3d8fcc3269a97c66029b918
https://git.kernel.org/pub/scm/linux/kernel/git/mnyman/xhci.git
for-usb-next)

BUG=b:181277177
TEST=No xhci related regressions

Signed-off-by: Charishma Gairuboyina <charishma1.gairuboyina@intel.com>
Change-Id: I727a19df30df72542d488677be9cce0f37cb1fc8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2727545
Tested-by: Charishma Gairuboyina <charishma1.gairuboyina@intel.corp-partner.google.com>
Tested-by: Rajat Jain <rajatja@google.com>
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Rajat Jain <rajatja@google.com>
Commit-Queue: Rajat Jain <rajatja@google.com>