CHROMIUM: Track adv state correctly when extended adv available
We have discovered that in some use cases, the HCI_LE_ADV will not
correctly reflect the advertising state when extended advertising is
available. Due to the logic from pre-extended implementation, when an
"enable advertising" command is received with enable=1, HCI_LE_ADV is
set to True, and if enable=0, HCI_LE_ADV is set to false. However, as
ext_adv allows multiple concurrent advertisements, this logic is now
incorrect. Consider the following sequence of bluez events:

1. RegisterAdvertisement(adv1)
2. RegisterAdvertisement(adv2)
3. UnregisterAdvertisement(adv2)

In step 3, HCI_LE_ADV is set to false even though adv1 is still active.
This patch does the following to improve correctness of the HCI_LE_ADV
flag:

* Since directed advertisements are not tracked via the
  hdev->adv_instances list, add a new hdev->ext_directed_advertising
  bool to track when directed advertising is occurring.
* Since extended advertising now has a new hci_le_ext_adv_term_evt event
  that fires when the controller decides to cancel an advertisement due
  to a timeout, a connect event to that adv's address, etc., allow
  HCI_LE_ADV to be updated here as well.
* Since hci_le_ext_adv_term_evt cancels an adv, we now remove this adv
  from hdev->adv_instances.

Note that in the third point here, we remove the cancelled adv from
hdev->adv_instances, but this patch does not handle the propagation to
userspace (and client). This patch simply syncs the kernel to the
controller's adv state, which is necessary to have an accurate count of
active advertisements.

The patch is tested with our QuickHealth and AdvHealth automated test
suites. Additionally, it is tested by performing the above steps and
then validating that the HCI_LE_ADV flag correctly represents the
controller state.

BUG=chromium:1163671
TEST=AdvHealth, QuickHealth on hatch

Signed-off-by: Daniel Winkler <danielwinkler@google.com>
Change-Id: Ia726eb7516c403a07e67fd30d5727720a78211fc
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2698813
Reviewed-by: Miao-chen Chou <mcchou@chromium.org>