FROMGIT: ath10k: skip the wait for completion to recovery in shutdown path
Currently in the shutdown callback we wait for recovery to complete
before freeing up the resources. This results in additional two seconds
delay during the shutdown and thereby increase the shutdown time.

As an attempt to take less time during shutdown, remove the wait for
recovery completion in the shutdown callback and added an API to freeing
the reosurces in which they were common for shutdown and removing
the module.

Tested-on: WCN3990 hw1.0 SNOC WLAN.HL.3.1-01040-QCAHLSWMTPLZ-1

Signed-off-by: Youghandhar Chintala <youghand@codeaurora.org>
Signed-off-by: Kalle Valo <kvalo@codeaurora.org>
Link: https://lore.kernel.org/r/20210223142908.23374-1-youghand@codeaurora.org
(cherry picked from commit 018e3fa8e7ff087629e5285cc14d1a6ca9bebc8b
 git://git.kernel.org/pub/scm/linux/kernel/git/kvalo/ath.git ath-next)

BUG=b:165343111
TEST=Build passes and device reboots quicker.

Disallow-Recycled-Builds: test-failures
Signed-off-by: Abhishek Kumar <kuabhs@chromium.org>
Change-Id: I6695a6266030f8687c4c8d0648fca5401d3bc6f7
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2728756
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Brian Norris <briannorris@chromium.org>