CHROMIUM: arm64: dts: qcom: Put the right panel in for limozeen
The first revision of limozeen (-rev4) had a different panel in
touchscreen devices.  That panel won't be on any real devices.  Let's
switch to the right panel.  This fixes a black blink at bootup on
-rev5+ touchscreen devices and also prevents them from running at the
wrong mode for a short time.  See b:172979067.

NOTE: as a side effect of this change we'll actually be running
"-rev4" touchscreen panels at the wrong mode for a little while and
_they'll_ get the blink.  That shouldn't be a big deal because:
* The "wrong" mode seems to work OK.
* The panels in -rev4 are pretty wonky anyway.  The EDID is slightly
  invalid and that makes the BIOS grumpy and it won't init the display
  (see b:169211952).
* No real hardware will have these displays.
* Eventually we'll fix b:172979067

BUG=b:181175014, b:172979067, b:168488026
TEST=No black blink on -rev5; yes black blink on -rev4.

Change-Id: I02cc4ad1a3fd9233e0de7d3684f99382296679e8
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2719129
Reviewed-by: Stephen Boyd <swboyd@chromium.org>