FROMLIST: media: mtk-vcodec: vdec: move stateful ops into their own file
We are planning to add support for stateless decoders to this driver.
Part of the driver will be shared between stateful and stateless
codecs, but a few ops need to be specialized for both. Extract the
stateful part of the driver and move it into its own file, accessible
through ops that the common driver parts can call.

This patch only moves code around and introduces a set of abstractions ;
the behavior of the driver should not be changed in any way.

Signed-off-by: Yunfei Dong <yunfei.dong@mediatek.com>
[acourbot: refactor, cleanup and split]
Co-developed-by: Alexandre Courbot <acourbot@chromium.org>
Signed-off-by: Alexandre Courbot <acourbot@chromium.org>
(am from https://patchwork.linuxtv.org/patch/71757/)
(also found at https://lore.kernel.org/r/20210226100148.1663389-2-acourbot@chromium.org)

BUG=b:158436469
TEST=Kernel builds for Asurada.

Change-Id: I81e6de43f00cf0e83c6464420dc9bf6e4532d6a9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2726275
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Tzung-Bi Shih <tzungbi@chromium.org>
Commit-Queue: Alexandre Courbot <acourbot@chromium.org>
Tested-by: Alexandre Courbot <acourbot@chromium.org>