CHROMIUM: thunderbolt: Enable retimer for IO before reading retimer NVM
In cases, where no devices are attached to a Type-C port, retimer
IO needs to be enabled before doing any IO with that retimer device.
This patch enables the retimer IO to be able to read the retimer
firmware.

BUG=b:181343919
TEST=Boot with no devices attached to Type-C port and read retimer
firmware from an active nvmem using below command on a Voxel EVT.
cat /sys/bus/thunderbolt/devices/0-0/0-0\:X.1/nvm_activeY/nvmem
The NVM read time should be close to the same amount of time,
observed when a TBT/USB4 device is connected.

TOREVERT=Once the upstream version of the patch lands in
maintainers tree, this patch should be reverted.
TOREVERT=b:180449562

Change-Id: I694c03857ce370c8ae431a4d3a206c92ee8c6671
Signed-off-by: Rajmohan Mani <rajmohan.mani@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2724836
Tested-by: Kranthi Kuntala <kranthi.kuntala@intel.corp-partner.google.com>
Reviewed-by: Prashant Malani <pmalani@chromium.org>