CHROMIUM: media: camx: Fix cache management, take 2
The previous attempt assumed that there would always be a device
attached to the buffer before the CPU accesses it, but it is not true.
In such case the current code will leave the cached kernel mapping
unflushed and its writeback would race with the CPU access via the
uncached mapping. Fix it by removing the conditional flushing and just
always doing it at allocation and free time.

BUG=b:168139312
BUG=b:177340920
TEST=Record video and take snapshot on both cameras on strongbad.

Signed-off-by: Tomasz Figa <tfiga@chromium.org>
Change-Id: Id95fca25e92bc1bd0dfe26886a50c11f46dba08a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2726795
Tested-by: Atanas Filipov <afilipov@mm-sol.com>
Reviewed-by: Sergey Senozhatsky <senozhatsky@chromium.org>