CHROMIUM: thunderbolt: Store size of active portion retimer NVMem device
Storing the size of the active portion of a retimer NVMem device is
useful, to decide when the retimer IO needs to be stopped, during
the read of active retimer NVMem device.

BUG=b:181343919
TEST=Ensure there are no regressions in use of Type-C based devices

TOREVERT=Once the upstream version of the patch lands in
maintainers tree, this patch should be reverted.
TOREVERT=b:180449562

Change-Id: Id43a5eddafb4579d730c5fa79548da968493e51c
Signed-off-by: Rajmohan Mani <rajmohan.mani@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2726667
Tested-by: Kranthi Kuntala <kranthi.kuntala@intel.corp-partner.google.com>
Reviewed-by: Prashant Malani <pmalani@chromium.org>