CHROMIUM: Disable COMPILE_TEST for camx driver
0-day reports lots of compile warnings and errors for this driver, too
many to list one by one. Fixing them would be a nightmare. The driver
is not ready for upstream integration, and is in poor shape. Attempts
to build it with W=1 report a large number of problems, some of them
pointing to real bugs. Various Makefiles point to include directories
which don't exist. One of the directories is orphan: No Makefile
points to it.

Worse, the driver does not even build with a recent version of gcc,
but fails with the following error.

drivers/media/platform/camx/cam_cdm/cam_cdm_util.c:311:2: error:
	positional initialization of field in 'struct' declared with
	'designated_init' attribute

Fixing all the reported problems is out of scope for the kernel team.
Disable build tests for this driver instead. That won't make a practical
difference for build tests running through CQ because those only test
allmodconfig and the driver is disabled for those anyway, but it will
limit error reporting by 0-day.

BUG=b:168139312
TEST=Test build with W=1

Change-Id: I53c3bfd2a8b6b2b7e931603bd9c0dbc8f9c36788
Disallow-Recycled-Builds: test-failures
Signed-off-by: Guenter Roeck <groeck@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2732085