CHROMIUM: iwl7000: mvm: fix the PROTOCOL_OFFLOAD_CMD for old devices
Old devices don't have the firmware TLV that advertises the
version of the command. In this case, iwl_fw_lookup_cmd_ver
returns the provided default. I mistakenly provided
IWL_FW_CMD_VER_UNKNOWN which is defined to 99 meaning that
in case we have an old firmware, we assumed that we support
the latest version of the command. This is a bug obviously.

Fix this by supplying 0 as the default value.

Change-Id: I5c1c8355560fac42340dc94a6a7fe101cad1bc77
Signed-off-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
iwl7000-tree: 0d1e657f38565ff2e28ce5324bfe9cf1d6379715