FROMGIT: thunderbolt: debugfs: Show all accessible dwords
Currently, when first failure occurs while reading of the block,
we stop reading the block and jump to the next capability.
This doesn't cover the case of block with "holes" of inaccessible
dwords, followed by accessible dwords.
This patch address this problem.
In case of failure while reading the complete block in one transaction,
(because of one or more dwords is inaccessible), we read the remaining
dwords of the block dword-by-dword, one dword per transaction,
till the end of the block.
By doing this, we handle the case of block with "holes" of inaccessible
dwords, followed by accessible dwords. The accessible dwords are shown
with the fields: <offset> <relative_offset> <cap_id> <vs_cap_id> <value>
E.g.:
0x01eb  236 0x05 0x06 0x0000d166
While the inaccesible dwords are shown as: <offset> <not accessible>
E.g.:
0x01ed <not accessible>

Signed-off-by: Gil Fine <gil.fine@intel.com>
Signed-off-by: Mika Westerberg <mika.westerberg@linux.intel.com>
(cherry picked from commit 3382665a5c5de586cd6e93f9e892527d0775a1bb
https://git.kernel.org/pub/scm/linux/kernel/git/westeri/thunderbolt.git next)

BUG=b:182237111
TEST=cat /sys/kernel/debug/thunderbolt/0-0/regs
Check for dwords within the previously "Not Accessible" region:
0x0600    0 0x05 0x07 0x00070500
0x0601    1 0x05 0x07 0x00800700
0x0602 <not accessible>
0x0603 <not accessible>
0x0604 <not accessible>
0x0605 <not accessible>
... (skipped lines)
0x063d <not accessible>
0x063e <not accessible>
0x063f   63 0x05 0x07 0x00000000
... (skipped lines)
0x067e  126 0x05 0x07 0x00000000
0x067f  127 0x05 0x07 0x00000000
0x0700    0 0x05 0x08 0x00080500

Signed-off-by: Casey Bowman <casey.g.bowman@intel.com>
Change-Id: Ic4968beea483e5f1ecb736022541bb750f4b01e6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2762777
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Prashant Malani <pmalani@chromium.org>