CHROMIUM: arm64: dts: qcom: Remove "dmic_clk_en" from coachz revs
It's the same in every revision and upstream has moved to the
coachz.dtsi file.

NOTE: it turns out that -r1 and -r2 of coachz are identical because
<https://crrev.com/c/2733863> ("CHROMIUM: arm64: dts: qcom: sc7180:
add dmic_clk_en back") wasn't the proper inverse of
<https://crrev.com/c/2596726> ("CHROMIUM: arm64: dts: qcom: sc7180:
remove dmic_clk_en"). We'll have to address this upstream since I
didn't notice and we've got both -r1 and -r2 there.

BUG=b:179637245, b:147399356
TEST=Use audio

Signed-off-by: Douglas Anderson <dianders@chromium.org>
Change-Id: Ia624a1902749ca95f7673b112bcf944bbdbd5e4a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2762002
Reviewed-by: Matthias Kaehlcke <mka@chromium.org>