CHROMIUM: thunderbolt: Cancel retimer scan work during retimer removal
When a system shutdown is initiated before the retimer scan work
is run, it is possible for the retimer scan work to be called,
after the driver is unloaded. This results in a kernel panic,
with a NULL pointer access. Fix this issue by checking and
canceling pending retimer scan work item for a given USB4 port
of the host router.

BUG=b:182828088
TEST=Run 2500 cold boot cycles on Voxel and ensure there are
no kernel panics.

TOREVERT=Once the upstream version of the patch lands in
maintainers tree, this patch should be reverted.
TOREVERT=b:180449562

Change-Id: Ic1948eeba686aae17455ff6ef5b49deeef8d670a
Signed-off-by: Srikanth Nandamuri <srikanth.nandamuri@intel.com>
Signed-off-by: Rajmohan Mani <rajmohan.mani@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2762252
Reviewed-by: Srikanth Nandamuri <srikanth.nandamuri@intel.corp-partner.google.com>
Reviewed-by: Prashant Malani <pmalani@chromium.org>
Tested-by: Srikanth Nandamuri <srikanth.nandamuri@intel.corp-partner.google.com>