BACKPORT: net: fix race between napi kthread mode and busy poll
Currently, napi_thread_wait() checks for NAPI_STATE_SCHED bit to
determine if the kthread owns this napi and could call napi->poll() on
it. However, if socket busy poll is enabled, it is possible that the
busy poll thread grabs this SCHED bit (after the previous napi->poll()
invokes napi_complete_done() and clears SCHED bit) and tries to poll
on the same napi. napi_disable() could grab the SCHED bit as well.
This patch tries to fix this race by adding a new bit
NAPI_STATE_SCHED_THREADED in napi->state. This bit gets set in
____napi_schedule() if the threaded mode is enabled, and gets cleared
in napi_complete_done(), and we only poll the napi in kthread if this
bit is set. This helps distinguish the ownership of the napi between
kthread and other scenarios and fixes the race issue.

Fixes: 29863d41bb6e ("net: implement threaded-able napi poll loop support")
Reported-by: Martin Zaharinov <micron10@gmail.com>
Suggested-by: Jakub Kicinski <kuba@kernel.org>
Signed-off-by: Wei Wang <weiwan@google.com>
Cc: Alexander Duyck <alexanderduyck@fb.com>
Cc: Eric Dumazet <edumazet@google.com>
Cc: Paolo Abeni <pabeni@redhat.com>
Cc: Hannes Frederic Sowa <hannes@stressinduktion.org>
Signed-off-by: Jakub Kicinski <kuba@kernel.org>
Signed-off-by: David S. Miller <davem@davemloft.net>

(cherry picked from commit cb038357937ee4f589aab2469ec3896dce90f317)

Conflicts:
   net/core/dev.c: conflict due to missing napi_prefer_busy_poll feature
   in kernel version 5.4

BUG=b:161825061
TEST=Verified throughput performance through patch build and can see
     improvement in throughput

Signed-off-by: Abhishek Kumar <kuabhs@chromium.org>
Change-Id: I4c7e2a77a47a699699bf5cdd3ccfda5fc55540cb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2859774
Reviewed-by: Brian Norris <briannorris@chromium.org>