CHROMIUM: thunderbolt: Remove external retimers during switch removal
When an external switch is removed, currently retimers that belong to
that switch are not removed. Enable this by allowing retimers to be
removed for external switches.

BUG=b:185219629
TEST=Ensure retimer sysfs entries that belong to external switches are
removed, once the switch is removed.
Login to voxel, connect Goshen Ridge (GR) to a Type-C port and connect
Belkin TBT dock to Port B of GR and check for GR retimer sysfs entries
under /sys/bus/thunderbolt/devices. Unplug GR and ensure that GR retimer
sysfs entries go away.

TOREVERT=Once the upstream version of the patch lands in
maintainers tree, this patch should be reverted.
TOREVERT=b:180449562

Change-Id: I2f850fede9469f78328b65b690b37589d213c131
Signed-off-by: Rajmohan Mani <rajmohan.mani@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2823232
Reviewed-by: Prashant Malani <pmalani@chromium.org>