CHROMIUM: drm/virtio: test potential race between device activation and getting capsets
**** Don't forward port or try upstreaming this patch, this only
     for internal testing!!! ***

v5.4 is going away anyways very soon.

There's a suspicion that flaky Crostini startups is caused by a
race between device activation and getting the capsets.  A
possible solution is triggering a virtio_cread(..) after
virtio_device_ready, which triggers a wait.  Probably not the
best solution, but should allow the lab to disapprove or approve
our theories.

If the theory is incorrect, this is just an harmless re-ordering
that should not be forward ported.

BUG=chromium:1171101
TEST=lab will test

Change-Id: Iba5780f0cdd3f8551432bb6685fd3e70ccf4f41f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2776131
Tested-by: Gurchetan Singh <gurchetansingh@chromium.org>
Reviewed-by: Ryan Neph <ryanneph@google.com>
Commit-Queue: Gurchetan Singh <gurchetansingh@chromium.org>