UPSTREAM: net: ipa: fix init header command validation
We use ipa_cmd_header_valid() to ensure certain values we will
program into hardware are within range, well in advance of when we
actually program them.  This way we avoid having to check for errors
when we actually program the hardware.

Unfortunately the dev_err() call for a bad offset value does not
supply the arguments to match the format specifiers properly.
Fix this.

There was also supposed to be a check to ensure the size to be
programmed fits in the field that holds it.  Add this missing check.

Rearrange the way we ensure the header table fits in overall IPA
memory range.

Finally, update ipa_cmd_table_valid() so the format of messages
printed for errors matches what's done in ipa_cmd_header_valid().

Signed-off-by: Alex Elder <elder@linaro.org>
Signed-off-by: David S. Miller <davem@davemloft.net>
(cherry picked from commit b4afd4b90a7cfe54c7cd9db49e3c36d552325eac)

BUG=b:175328613
TEST=Modem functional, No IPA hang in suspend stress test

Signed-off-by: Sujit Kautkar <sujitka@chromium.org>
Change-Id: I085a58ff47668b457dfd74388d11e1ac92208b39
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2918736
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Douglas Anderson <dianders@chromium.org>