CHROMIUM: MALI: Fix NULL pointer access issue on context create failure
(GPUCORE-27491)
To fix GPUCORE-26305, a workaround was added in the kernel driver code,
later a fix was committed, which made that workaround redundant.
But the workaround was not removed and its location in the code was
changed such that it also got exercised when the context initialization
itself failed.
This resulted in a NULL pointer de-reference issue when there is an
error in the context creation sequence, before the point where the
'ctx_list_entry[]' list heads are initialized.

This commit fixes the NULL pointer access issue by reverting that
workaround.

BUG=b:183068153
TEST=test_that ${IP} graphics_Stress.restart --iterations 10

Change-Id: Ifd131ac1da564b7f11cceb823d1cbec66f3a9475
Signed-off-by: Fei Shao <fshao@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2791144
Reviewed-by: Renato Grottesi <renato.grottesi@arm.corp-partner.google.com>