UPSTREAM: drm/msm/dsi: stop passing src_pll_id to the phy_enable call
Phy driver already knows the source PLL id basing on the set usecase and
the current PLL id. Stop passing it to the phy_enable call. As a
reminder, dsi manager will always use DSI 0 as a clock master in a slave
mode, so PLL 0 is always a clocksource for DSI 0 and it is always a
clocksource for DSI 1 too unless DSI 1 is used in the standalone mode.

Signed-off-by: Dmitry Baryshkov <dmitry.baryshkov@linaro.org>
Tested-by: Stephen Boyd <swboyd@chromium.org> # on sc7180 lazor
Reviewed-by: Abhinav Kumar <abhinavk@codeaurora.org>
Link: https://lore.kernel.org/r/20210331105735.3690009-25-dmitry.baryshkov@linaro.org
Signed-off-by: Rob Clark <robdclark@chromium.org>
(cherry picked from commit 36c5dde5fdf028daf005197a4dd08b196784bf9e)

BUG=b:162602229
TEST=Hang DPU on Lazor; see devcoredump

Signed-off-by: Stephen Boyd <swboyd@chromium.org>
Change-Id: Ie2c16dcc83d31a70f88700cb2af774493d55e854
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2938246
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Rob Clark <robdclark@chromium.org>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>